// src/app/state/items/actions/items.types.ts
export enum ItemsActionTypes {
    LOAD_ITEMS = '[items] load items',
    LOAD_ITEMS_SUCCESS = '[items] load items success',
    LOAD_ITEMS_ERROR = '[items] load items error',
    ITEM_FILTER_PROPS =  '[items] filter props',

    LOAD_SINGLE_ITEM = '[items] load single item',
    LOAD_SINGLE_ITEM_SUCCESS = '[items] load single item success',
    LOAD_SINGLE_ITEM_ERROR = '[items] load single item error',

    LOAD_ITEM_OWNER_DETAILS = '[items] load item owner details',
    LOAD_ITEM_OWNER_DETAILS_SUCCESS = '[items] load item owner details success',
    LOAD_ITEM_OWNER_DETAILS_ERROR = '[items] load item owner details error',

   
}